requirements-dev.txt
pytest>=8.0
pyyaml>=6.0
